Move constants to root package
