+++ src/main/java/org/stratumproject/fabric/tna/behaviour/pipeliner/FabricPipeliner.java
@@ -64,15 +64,15 @@ import java.util.stream.Collectors;
 
 import static java.lang.String.format;
 import static org.slf4j.LoggerFactory.getLogger;
-import static org.stratumproject.fabric.tna.behaviour.Constants.DEFAULT_VLAN;
-import static org.stratumproject.fabric.tna.behaviour.Constants.FWD_IPV4_ROUTING;
-import static org.stratumproject.fabric.tna.behaviour.Constants.FWD_MPLS;
-import static org.stratumproject.fabric.tna.behaviour.Constants.ONE;
-import static org.stratumproject.fabric.tna.behaviour.Constants.PORT_TYPE_INTERNAL;
-import static org.stratumproject.fabric.tna.behaviour.Constants.RECIRC_PORTS;
-import static org.stratumproject.fabric.tna.behaviour.Constants.V1MODEL_RECIRC_PORT;
-import static org.stratumproject.fabric.tna.behaviour.Constants.ZERO;
-import static org.stratumproject.fabric.tna.behaviour.Constants.PKT_IN_MIRROR_SESSION_ID;
+import static org.stratumproject.fabric.tna.Constants.DEFAULT_VLAN;
+import static org.stratumproject.fabric.tna.Constants.FWD_IPV4_ROUTING;
+import static org.stratumproject.fabric.tna.Constants.FWD_MPLS;
+import static org.stratumproject.fabric.tna.Constants.ONE;
+import static org.stratumproject.fabric.tna.Constants.PORT_TYPE_INTERNAL;
+import static org.stratumproject.fabric.tna.Constants.RECIRC_PORTS;
+import static org.stratumproject.fabric.tna.Constants.V1MODEL_RECIRC_PORT;
+import static org.stratumproject.fabric.tna.Constants.ZERO;
+import static org.stratumproject.fabric.tna.Constants.PKT_IN_MIRROR_SESSION_ID;
 import static org.stratumproject.fabric.tna.behaviour.FabricUtils.KRYO;
 
 import static org.stratumproject.fabric.tna.behaviour.FabricUtils.outputPort;
