+++ src/main/java/org/stratumproject/fabric/tna/behaviour/pipeliner/FilteringObjectiveTranslator.java
@@ -44,14 +44,14 @@ import static org.onosproject.segmentrouting.metadata.SRObjectiveMetadata.CLEANU
 import static org.onosproject.segmentrouting.metadata.SRObjectiveMetadata.INTERFACE_CONFIG_UPDATE;
 import static org.onosproject.segmentrouting.metadata.SRObjectiveMetadata.isSrMetadataSet;
 import static org.onosproject.segmentrouting.metadata.SRObjectiveMetadata.isValidSrMetadata;
-import static org.stratumproject.fabric.tna.behaviour.Constants.ETH_TYPE_EXACT_MASK;
-import static org.stratumproject.fabric.tna.behaviour.Constants.FWD_IPV4_ROUTING;
-import static org.stratumproject.fabric.tna.behaviour.Constants.FWD_IPV6_ROUTING;
-import static org.stratumproject.fabric.tna.behaviour.Constants.FWD_MPLS;
-import static org.stratumproject.fabric.tna.behaviour.Constants.ONE;
-import static org.stratumproject.fabric.tna.behaviour.Constants.PORT_TYPE_EDGE;
-import static org.stratumproject.fabric.tna.behaviour.Constants.PORT_TYPE_INFRA;
-import static org.stratumproject.fabric.tna.behaviour.Constants.ZERO;
+import static org.stratumproject.fabric.tna.Constants.ETH_TYPE_EXACT_MASK;
+import static org.stratumproject.fabric.tna.Constants.FWD_IPV4_ROUTING;
+import static org.stratumproject.fabric.tna.Constants.FWD_IPV6_ROUTING;
+import static org.stratumproject.fabric.tna.Constants.FWD_MPLS;
+import static org.stratumproject.fabric.tna.Constants.ONE;
+import static org.stratumproject.fabric.tna.Constants.PORT_TYPE_EDGE;
+import static org.stratumproject.fabric.tna.Constants.PORT_TYPE_INFRA;
+import static org.stratumproject.fabric.tna.Constants.ZERO;
 import static org.stratumproject.fabric.tna.behaviour.FabricUtils.l2InstructionOrFail;
 import static org.stratumproject.fabric.tna.behaviour.FabricUtils.criterion;
 import static org.stratumproject.fabric.tna.behaviour.FabricUtils.l2Instruction;
