+++ src/main/java/org/stratumproject/fabric/tna/behaviour/pipeliner/ForwardingObjectiveTranslator.java
@@ -37,9 +37,9 @@ import static java.lang.String.format;
 import static org.onosproject.segmentrouting.metadata.SRObjectiveMetadata.PAIR_PORT;
 import static org.onosproject.segmentrouting.metadata.SRObjectiveMetadata.isSrMetadataSet;
 import static org.onosproject.segmentrouting.metadata.SRObjectiveMetadata.isValidSrMetadata;
-import static org.stratumproject.fabric.tna.behaviour.Constants.PORT_TYPE_EDGE;
-import static org.stratumproject.fabric.tna.behaviour.Constants.PORT_TYPE_INFRA;
-import static org.stratumproject.fabric.tna.behaviour.Constants.PORT_TYPE_MASK;
+import static org.stratumproject.fabric.tna.Constants.PORT_TYPE_EDGE;
+import static org.stratumproject.fabric.tna.Constants.PORT_TYPE_INFRA;
+import static org.stratumproject.fabric.tna.Constants.PORT_TYPE_MASK;
 import static org.stratumproject.fabric.tna.behaviour.FabricUtils.criterionNotNull;
 import static org.stratumproject.fabric.tna.behaviour.FabricUtils.outputPort;
 import static org.stratumproject.fabric.tna.behaviour.FabricUtils.portType;
