+++ src/test/java/org/stratumproject/fabric/tna/behaviour/FabricInterpreterTest.java
@@ -41,7 +41,7 @@ import static org.easymock.EasyMock.createNiceMock;
 import static org.easymock.EasyMock.expect;
 import static org.easymock.EasyMock.replay;
 import static org.junit.Assert.assertEquals;
-import static org.stratumproject.fabric.tna.behaviour.Constants.QUEUE_ID_SYSTEM;
+import static org.stratumproject.fabric.tna.Constants.QUEUE_ID_SYSTEM;
 
 /**
  * Test for fabric interpreter.
