+++ src/test/java/org/stratumproject/fabric/tna/slicing/SlicingManagerTest.java
@@ -56,9 +56,9 @@ import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 import static org.onlab.junit.TestTools.assertAfter;
-import static org.stratumproject.fabric.tna.behaviour.Constants.COLOR_GREEN;
-import static org.stratumproject.fabric.tna.behaviour.Constants.COLOR_RED;
-import static org.stratumproject.fabric.tna.behaviour.Constants.TNA;
+import static org.stratumproject.fabric.tna.Constants.COLOR_GREEN;
+import static org.stratumproject.fabric.tna.Constants.COLOR_RED;
+import static org.stratumproject.fabric.tna.Constants.TNA;
 import static org.stratumproject.fabric.tna.behaviour.FabricUtils.sliceTcConcat;
 import static org.stratumproject.fabric.tna.behaviour.P4InfoConstants.FABRIC_INGRESS_QOS_DEFAULT_TC;
 import static org.stratumproject.fabric.tna.behaviour.P4InfoConstants.FABRIC_INGRESS_QOS_QUEUES;
